+++ src/Game/GameObjects/Eagle.h
@@ -21,7 +21,7 @@ public:
 
     Eagle(const glm::vec2& position, const glm::vec2& size, const float rotation, const float layer);
     virtual void render() const override;
-    virtual void update(uint64_t delta) override;
+    virtual void update(double delta) override;
 
 private:
     std::array<std::shared_ptr<RenderEngine::Sprite>, 2> m_sprites;
